Keep symlink containment logic at the tool boundary
The first draft fixed the reported bug but also pushed canonical-path helpers down into common.ts, which made the PR broader than necessary and muddied the design boundary. This follow-up keeps the symlink-aware containment check local to the MCP tool, matching the existing imageDir contract and leaving shared download helpers unchanged.

Constraint: The tool edge is the only layer that knows the configured imageDir boundary
Rejected: Re-export path helpers from common.ts | unnecessary API surface for a single call site
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep imageDir boundary enforcement where imageDir is available; shared helpers should stay boundary-agnostic
Tested: ./node_modules/.bin/vitest run src/tests/path-validation.test.ts; ./node_modules/.bin/tsc --noEmit; ./node_modules/.bin/eslint src/mcp/tools/download-figma-images-tool.ts src/tests/path-validation.test.ts
Not-tested: Full pnpm test suite; cross-platform symlink behavior on Windows
